+++ packages/list-module/__tests__/elem-to-html.test.ts
@@ -10,11 +10,11 @@ import listItemToHtmlConf from '../src/module/elem-to-html'
 describe('module elem-to-html', () => {
   const childrenHtml = '<span>hello</span>'
 
-  const orderedElem1 = { type: 'list2-item', ordered: true, children: [{ text: '' }] }
-  const orderedElem2 = { type: 'list2-item', ordered: true, children: [{ text: '' }] }
-  const unOrderedItem1 = { type: 'list2-item', children: [{ text: '' }] }
-  const unOrderedItem2 = { type: 'list2-item', children: [{ text: '' }] }
-  const unOrderedItem21 = { type: 'list2-item', level: 1, children: [{ text: '' }] }
+  const orderedElem1 = { type: 'list-item', ordered: true, children: [{ text: '' }] }
+  const orderedElem2 = { type: 'list-item', ordered: true, children: [{ text: '' }] }
+  const unOrderedItem1 = { type: 'list-item', children: [{ text: '' }] }
+  const unOrderedItem2 = { type: 'list-item', children: [{ text: '' }] }
+  const unOrderedItem21 = { type: 'list-item', level: 1, children: [{ text: '' }] }
 
   const editor = createEditor({
     content: [orderedElem1, orderedElem2, unOrderedItem1, unOrderedItem2, unOrderedItem21],
@@ -28,7 +28,7 @@ describe('module elem-to-html', () => {
   ELEM_TO_EDITOR.set(unOrderedItem21, editor)
 
   test(`toHtml conf type`, () => {
-    expect(listItemToHtmlConf.type).toBe('list2-item')
+    expect(listItemToHtmlConf.type).toBe('list-item')
   })
 
   test(`ordered item toHtml`, () => {