+++ packages/list-module/__tests__/render-elem.test.ts
@@ -4,22 +4,22 @@
  */
 
 import createEditor from '../../../tests/utils/create-editor'
-import renderList2ItemConf from '../src/module/render-elem'
+import renderListItemConf from '../src/module/render-elem'
 
 describe('list module - render elem', () => {
-  const unOrderedItem = { type: 'list2-item', children: [{ text: '' }] }
-  const orderedItem = { type: 'list2-item', ordered: true, children: [{ text: '' }] }
-  const leveledItem = { type: 'list2-item', level: 3, children: [{ text: '' }] }
+  const unOrderedItem = { type: 'list-item', children: [{ text: '' }] }
+  const orderedItem = { type: 'list-item', ordered: true, children: [{ text: '' }] }
+  const leveledItem = { type: 'list-item', level: 3, children: [{ text: '' }] }
   const editor = createEditor({
     content: [unOrderedItem, orderedItem, leveledItem],
   })
 
   it('render conf type', () => {
-    expect(renderList2ItemConf.type).toBe('list2-item')
+    expect(renderListItemConf.type).toBe('list-item')
   })
 
   it('render ordered list item elem', () => {
-    const vnode: any = renderList2ItemConf.renderElem(orderedItem, null, editor)
+    const vnode: any = renderListItemConf.renderElem(orderedItem, null, editor)
     expect(vnode.sel).toBe('div') // render-elem 使用 <div> 模拟 <li>
 
     const prefixVnode = vnode.children[0] || {}
@@ -27,7 +27,7 @@ describe('list module - render elem', () => {
   })
 
   it('render unOrdered list item elem', () => {
-    const vnode: any = renderList2ItemConf.renderElem(unOrderedItem, null, editor)
+    const vnode: any = renderListItemConf.renderElem(unOrderedItem, null, editor)
     expect(vnode.sel).toBe('div') // render-elem 使用 <div> 模拟 <li>
 
     const prefixVnode = vnode.children[0] || {}
@@ -35,7 +35,7 @@ describe('list module - render elem', () => {
   })
 
   it('render leveled list item elem', () => {
-    const vnode: any = renderList2ItemConf.renderElem(leveledItem, null, editor)
+    const vnode: any = renderListItemConf.renderElem(leveledItem, null, editor)
     const style = vnode.data.style
     expect(style).toEqual({ margin: '5px 0 5px 60px' }) // margin-left 60px
   })