+++ packages/list-module/src/module/custom-types.ts
@@ -1,5 +1,5 @@
 /**
- * @description list2 element
+ * @description list element
  * @author wangfupeng
  */
 
@@ -7,8 +7,8 @@ import { Text } from 'slate'
 
 //【注意】需要把自定义的 Element 引入到最外层的 custom-types.d.ts
 
-export type List2ItemElement = {
-  type: 'list2-item'
+export type ListItemElement = {
+  type: 'list-item'
   ordered: boolean // 有序/无序
   level: number // 层级：0 1 2 ...
   children: Text[]