+++ packages/list-module/src/module/menu/index.ts
@@ -7,14 +7,14 @@ import BulletedListMenu from './BulletedListMenu'
 import NumberedListMenu from './NumberedListMenu'
 
 export const bulletedListMenuConf = {
-  key: 'bulletedList2',
+  key: 'bulletedList',
   factory() {
     return new BulletedListMenu()
   },
 }
 
 export const numberedListMenuConf = {
-  key: 'numberedList2',
+  key: 'numberedList',
   factory() {
     return new NumberedListMenu()
   },