+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - [code-highlight](./packages/code-highlight/CHANGELOG.md)
 - [core](./packages/core/CHANGELOG.md)
 - [editor](./packages/editor/CHANGELOG.md)
-- [list2-module](./packages/list2-module/CHANGELOG.md)
+- [list-module](./packages/list-module/CHANGELOG.md)
 - [table-module](./packages/table-module/CHANGELOG.md)
 - [upload-image-module](./packages/upload-image-module/CHANGELOG.md)
 - [video-module](./packages/video-module/CHANGELOG.md)